requirements.txt
requests
faker
ldap3
PyJWT
